+++ Wrappers/Python/cil/plugins/astra/operators/ProjectionOperator.py
@@ -14,10 +14,9 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-
 from cil.framework import DataOrder
 from cil.optimisation.operators import LinearOperator, ChannelwiseOperator
-from cil.framework.BlockGeometry import  BlockGeometry
+from cil.framework.BlockGeometry import BlockGeometry
 from cil.optimisation.operators import BlockOperator
 from cil.plugins.astra.operators import AstraProjector3D
 from cil.plugins.astra.operators import AstraProjector2D
@@ -65,10 +64,11 @@ class ProjectionOperator(LinearOperator):
             return BlockOperator(*K)
         else:
             logging.info("Standard Operator is returned.")
-            return super(ProjectionOperator, cls).__new__(ProjectionOperator_ag)
+            return super(ProjectionOperator,
+                         cls).__new__(ProjectionOperator_ag)
+
 
 class ProjectionOperator_ag(ProjectionOperator):
-
     """
     ProjectionOperator configures and calls appropriate ASTRA Projectors for your dataset.
 
@@ -96,18 +96,25 @@ class ProjectionOperator_ag(ProjectionOperator):
     For multichannel data the ProjectionOperator will broadcast across all channels.
     """
 
-    def __init__(self, image_geometry=None, acquisition_geometry=None, device='gpu'):
-        
+    def __init__(self,
+                 image_geometry=None,
+                 acquisition_geometry=None,
+                 device='gpu'):
+
         if acquisition_geometry is None:
-            raise TypeError("Please specify an acquisition_geometry to configure this operator")
+            raise TypeError(
+                "Please specify an acquisition_geometry to configure this operator"
+            )
 
         if image_geometry is None:
             image_geometry = acquisition_geometry.get_ImageGeometry()
 
-        super(ProjectionOperator_ag, self).__init__(domain_geometry=image_geometry, range_geometry=acquisition_geometry)
+        super(ProjectionOperator_ag,
+              self).__init__(domain_geometry=image_geometry,
+                             range_geometry=acquisition_geometry)
 
         DataOrder.check_order_for_engine('astra', image_geometry)
-        DataOrder.check_order_for_engine('astra', acquisition_geometry) 
+        DataOrder.check_order_for_engine('astra', acquisition_geometry)
 
         self.volume_geometry = image_geometry
         self.sinogram_geometry = acquisition_geometry
@@ -116,14 +123,18 @@ class ProjectionOperator_ag(ProjectionOperator):
         volume_geometry_sc = image_geometry.get_slice(channel=0)
 
         if device == 'gpu':
-            operator = AstraProjector3D(volume_geometry_sc, sinogram_geometry_sc)
+            operator = AstraProjector3D(volume_geometry_sc,
+                                        sinogram_geometry_sc)
         elif self.sinogram_geometry.dimension == '2D':
-            operator = AstraProjector2D(volume_geometry_sc, sinogram_geometry_sc,  device=device)
+            operator = AstraProjector2D(volume_geometry_sc,
+                                        sinogram_geometry_sc,
+                                        device=device)
         else:
             raise NotImplementedError("Cannot process 3D data without a GPU")
 
-        if acquisition_geometry.channels > 1: 
-            operator_full = ChannelwiseOperator(operator, self.sinogram_geometry.channels, dimension='prepend')
+        if acquisition_geometry.channels > 1:
+            operator_full = ChannelwiseOperator(
+                operator, self.sinogram_geometry.channels, dimension='prepend')
             self.operator = operator_full
         else:
             self.operator = operator
@@ -146,7 +157,7 @@ class ProjectionOperator_ag(ProjectionOperator):
         '''
 
         return self.operator.direct(IM, out=out)
-    
+
     def adjoint(self, DATA, out=None):
         '''Applies the adjoint of the operator, i.e. the backward projection.
 
@@ -164,12 +175,12 @@ class ProjectionOperator_ag(ProjectionOperator):
             The processed data. Suppressed if `out` is passed
         '''
         return self.operator.adjoint(DATA, out=out)
-    
+
     def calculate_norm(self):
-        return self.operator.norm()    
+        return self.operator.norm()
 
     def domain_geometry(self):
         return self.volume_geometry
-    
+
     def range_geometry(self):
         return self.sinogram_geometry